requirements.txt
werkzeug==2.3.7
flask
flask_sqlalchemy
flask_login
flask_wtf
WTForms
wtforms-alchemy
SQLAlchemy
app
email_validator
python-dotenv
pyotp
flask-qrcode
MarkupSafe
bcrypt
rsa
flask-talisman
flask-testing
pytest-flask
beautifulsoup4